Moved line/width file and changed ohlc shader to rect shader
