+++ packages/d3fc-webgl/index.js
@@ -11,7 +11,7 @@ export { default as squarePointShader } from './src/shaders/point/square/shader'
 export { default as pointFill } from './src/shaders/point/fill';
 export { default as pointStroke } from './src/shaders/point/stroke';
 export { default as pointAntiAlias } from './src/shaders/point/antiAlias';
-export { default as lineWidth } from './src/shaders/line/width';
+export { default as lineWidth } from './src/shaders/lineWidth';
 export { default as barFill } from './src/shaders/bar/fill';
 
 export { default as glPoint } from './src/series/glPoint';
