+++ packages/d3fc-webgl/src/shaders/fragmentShaderSnippets.js
@@ -24,7 +24,7 @@ export const triangle = {
     }`
 }
 
-export const ohlc = {
+export const rect = {
     header: `varying float vColorIndicator;`,
     body: `gl_FragColor = vec4(0.4, 0.8, 0, 1);
     if (vColorIndicator < 0.0) {
