+++ packages/d3fc-webgl/src/shaders/lineWidth.js
